feat: Deploy Docker Secrets to the Docker Swarm
Docker Swarm mode does not yet understand the Docker Compose `secrets`
config in the same way that `docker-compose` does.  Docker Swarm will
not create secrets listed in a stack file (docker-compose.yaml).  It
will however, happily link services to existing, "external" secrets and
this is the purpose of this change: to create external secrets in the
Docker Swarm.  This is done in a step before the stack is deployed.