feat: Create external, primed docker volumes for stacks
By setting Values.primed_volumes with name, target and path properties,
stacks can define external docker volumes and copy files into them.

The name property is the name of the volume, the target property is the
mount point in the container and the path property is a file or
directory, under the stack remote path, to copy into the volume (to
"prime" it).

Note the the priming step creates a file ".done" in the "path".  This is
used to avoid repeated priming of the same volume.  Remove ".done" when
you want to force the volume to be re-primed.